+++ test/bolt_sips/protocol_test.exs
@@ -7,23 +7,27 @@ defmodule Bolt.Sips.ProtocolTest do
   # But works fine from Bolt.Sips.transaction
 
   test "connect/1 - disconnect/1 successful" do
-    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _} = conn_data} = Protocol.connect([])
+    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _, configuration: _} = conn_data} =
+             Protocol.connect([])
+
     assert :ok = Protocol.disconnect(:stop, conn_data)
   end
 
   test "checkout/1 successful" do
-    {:ok, %Protocol.ConnData{sock: _, bolt_version: _} = conn_data} = Protocol.connect([])
+    {:ok, %Protocol.ConnData{sock: _, bolt_version: _, configuration: _} = conn_data} =
+      Protocol.connect([])
 
-    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _} = conn_data} =
+    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _, configuration: _} = conn_data} =
              Protocol.checkout(conn_data)
 
     :ok = Protocol.disconnect(:stop, conn_data)
   end
 
   test "checkin/1 successful" do
-    {:ok, %Protocol.ConnData{sock: _, bolt_version: _} = conn_data} = Protocol.connect([])
+    {:ok, %Protocol.ConnData{sock: _, bolt_version: _, configuration: _} = conn_data} =
+      Protocol.connect([])
 
-    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _} = conn_data} =
+    assert {:ok, %Protocol.ConnData{sock: _, bolt_version: _, configuration: _} = conn_data} =
              Protocol.checkin(conn_data)
 
     :ok = Protocol.disconnect(:stop, conn_data)