+++ test/config_test.exs
@@ -13,39 +13,41 @@ defmodule Config.Test do
   ]
 
   @light_tls_config [
-    url: 'bolt://xmas:Kr1ngl3@hobby-happyHoHoHo.dbs.graphenedb.com:24786',
+    url: "bolt://xmas:Kr1ngl3@hobby-happyHoHoHo.dbs.graphenedb.com:24786",
     ssl: true
   ]
 
   @mixed_config [
-    hostname: '127.0.0.1',
+    hostname: "127.0.0.1",
     port: 0,
     url: @graphenedb_like_url
   ]
 
   @basic_config [
-    hostname: 'hobby',
+    hostname: "hobby",
     basic_auth: [username: "neo4j", password: "neo4j"],
     port: 1234,
     pool_size: 15,
-    max_overflow: 3
+    max_overflow: 3,
+    prefix: :default
   ]
 
   test "parsing the host and the port, from a url string config parameter" do
     config = Utils.default_config(@basic_tls_config)
 
     assert config[:url] == @graphenedb_like_url
-    assert config[:hostname] == 'hobby-happyHoHoHo.dbs.graphenedb.com'
+    assert config[:hostname] == "hobby-happyHoHoHo.dbs.graphenedb.com"
     assert config[:basic_auth] == [username: "xmas", password: "Kr1ngl3"]
     assert config[:port] == 24786
     assert config[:ssl] == true
+    assert config[:prefix] == :default
   end
 
   test "url string in config overides the :hostname and the :port" do
     config = Utils.default_config(@mixed_config)
 
     assert config[:url] == @graphenedb_like_url
-    assert config[:hostname] == 'hobby-happyHoHoHo.dbs.graphenedb.com'
+    assert config[:hostname] == "hobby-happyHoHoHo.dbs.graphenedb.com"
     assert config[:port] == 24786
   end
 
@@ -53,7 +55,7 @@ defmodule Config.Test do
     config = Utils.default_config(@basic_config)
 
     assert config[:url] == nil
-    assert config[:hostname] == 'hobby'
+    assert config[:hostname] == "hobby"
     assert config[:basic_auth] == [username: "neo4j", password: "neo4j"]
     assert config[:port] == 1234
     assert config[:ssl] == false
@@ -63,7 +65,7 @@ defmodule Config.Test do
     config = Utils.default_config(@light_tls_config)
 
     assert config[:url] != nil
-    assert config[:hostname] == 'hobby-happyHoHoHo.dbs.graphenedb.com'
+    assert config[:hostname] == "hobby-happyHoHoHo.dbs.graphenedb.com"
     assert config[:basic_auth] == [username: "xmas", password: "Kr1ngl3"]
     assert config[:port] == 24786
     assert config[:ssl] == true
@@ -72,15 +74,28 @@ defmodule Config.Test do
   test "standard Bolt.Sips default configuration" do
     config = Utils.default_config([])
 
-    assert config[:hostname] == 'localhost'
+    assert config[:hostname] == "localhost"
     assert config[:port] == 7687
     assert config[:ssl] == false
   end
 
-  test "url in configuration and :port" do
-    config = Utils.default_config(url: 'example.com', port: 123)
+  test "invalid url in configuration and explicit :port" do
+    config = Utils.default_config(url: "example.com", port: 123)
 
-    assert config[:hostname] == 'example.com'
-    assert config[:port] == 123
+    assert config[:hostname] == nil
+    assert config[:port] == 7687
+  end
+
+  test "url with routing context" do
+    config =
+      [url: "bolt+routing://neo4j:password@neo01.graph.example.com:123456?policy=europe"]
+      |> Utils.default_config()
+
+    assert config[:routing_context] == %{"policy" => "europe"}
+    assert config[:schema] == "bolt+routing"
+    assert config[:hostname] == "neo01.graph.example.com"
+    assert config[:port] == 123_456
+    assert config[:query] == "policy=europe"
+    assert config[:basic_auth] == [username: "neo4j", password: "password"]
   end
 end