+++ test/errors_test.exs
@@ -13,27 +13,25 @@ defmodule ErrorsTest do
   }
 
   test "create a node using SET properties and a simple map" do
-    r =
+    %Bolt.Sips.Response{stats: stats, type: type} =
       Bolt.Sips.query!(Bolt.Sips.conn(), "CREATE (report:Report) SET report = {props}", %{
         props: @simple_map
       })
 
-    assert r == %{
-             stats: %{"labels-added" => 1, "nodes-created" => 1, "properties-set" => 2},
-             type: "w"
-           }
+    assert %{"labels-added" => 1, "nodes-created" => 1, "properties-set" => 2} == stats
+    assert "w" == type
   end
 
   test "exception when creating a node using SET properties with a nested map" do
-    assert_raise Bolt.Sips.Exception,
-                 "Property values can only be of primitive types or arrays thereof",
-                 fn ->
-                   Bolt.Sips.query!(
-                     Bolt.Sips.conn(),
-                     "CREATE (report:Report) SET report = {props}",
-                     %{props: @nested_map}
-                   )
-                 end
+    err = "Property values can only be of primitive types or arrays thereof"
+
+    assert_raise Bolt.Sips.Exception, err, fn ->
+      Bolt.Sips.query!(
+        Bolt.Sips.conn(),
+        "CREATE (report:Report) SET report = {props}",
+        %{props: @nested_map}
+      )
+    end
   end
 
   test "exception when creating a node using SET properties with a list" do