+++ test/invalid_param_type_test.exs
@@ -2,6 +2,7 @@ defmodule Bolt.Sips.InvalidParamType.Test do
   use ExUnit.Case
 
   setup_all do
+    Bolt.Sips.ConnectionSupervisor.connections()
     {:ok, [conn: Bolt.Sips.conn()]}
   end
 
@@ -12,7 +13,7 @@ defmodule Bolt.Sips.InvalidParamType.Test do
       MATCH (n:Person {invalid: {an_elixir_datetime}}) RETURN TRUE
     """
 
-    {:error, [code: :failure, message: message]} =
+    {:error, %Bolt.Sips.Error{message: message}} =
       Bolt.Sips.query(conn, cypher, %{an_elixir_tuple: {:not, :valid}})
 
     assert String.match?(message, ~r/unable to encode value: {:not, :valid}/i)