+++ test/support/conn_case.ex
@@ -2,6 +2,7 @@ defmodule Bolt.Sips.ConnCase do
   use ExUnit.CaseTemplate
 
   setup_all do
+    Bolt.Sips.start_link(Application.get_env(:bolt_sips, Bolt))
     conn = Bolt.Sips.conn()
 
     on_exit(fn ->