+++ test/test_support.exs
@@ -1,0 +1,3 @@
+defmodule TestSupport do
+  @moduledoc false
+end