+++ test/transaction_test.exs
@@ -1,6 +1,8 @@
 defmodule Transaction.Test do
   use ExUnit.Case, async: true
 
+  alias Bolt.Sips.Response
+
   setup do
     {:ok, [main_conn: Bolt.Sips.conn()]}
   end
@@ -9,7 +11,7 @@ defmodule Transaction.Test do
     Bolt.Sips.transaction(main_conn, fn conn ->
       book =
         Bolt.Sips.query!(conn, "CREATE (b:Book {title: \"The Game Of Trolls\"}) return b")
-        |> List.first()
+        |> Response.first()
 
       assert %{"b" => g_o_t} = book
       assert g_o_t.properties["title"] == "The Game Of Trolls"
@@ -17,7 +19,7 @@ defmodule Transaction.Test do
     end)
 
     books = Bolt.Sips.query!(main_conn, "MATCH (b:Book {title: \"The Game Of Trolls\"}) return b")
-    assert length(books) == 0
+    assert Enum.count(books) == 0
   end
 
   ###
@@ -32,18 +34,23 @@ defmodule Transaction.Test do
   test "rollback statements in transaction", %{main_conn: main_conn} do
     try do
       # In case there's already a copy in our DB, count them...
-      {:ok, [result]} = Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
+      {:ok, %Response{results: [result]}} =
+        Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
+
       original_count = result["count(x)"]
 
       Bolt.Sips.transaction(main_conn, fn conn ->
-        book =
-          Bolt.Sips.query(conn, "CREATE (x:XactRollback {title:\"The Game Of Trolls\"}) return x")
+        assert {:ok, %Response{results: [row]}} =
+                 Bolt.Sips.query(
+                   conn,
+                   "CREATE (x:XactRollback {title:\"The Game Of Trolls\"}) return x"
+                 )
 
-        assert {:ok, [row]} = book
         assert row["x"].properties["title"] == "The Game Of Trolls"
 
         # Original connection (outside the transaction) should not see this node.
-        {:ok, [result]} = Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
+        assert {:ok, %Response{results: [result]}} =
+                 Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
 
         assert result["count(x)"] == original_count,
                "Main connection should not be able to see transactional change"
@@ -52,7 +59,9 @@ defmodule Transaction.Test do
       end)
 
       # Original connection should still not see this node committed.
-      {:ok, [result]} = Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
+      assert {:ok, %Response{results: [result]}} =
+               Bolt.Sips.query(main_conn, "MATCH (x:XactRollback) RETURN count(x)")
+
       assert result["count(x)"] == original_count
     after
       # Delete all XactRollback nodes in case the rollback() didn't work!
@@ -63,12 +72,21 @@ defmodule Transaction.Test do
   test "commit statements in transaction", %{main_conn: main_conn} do
     try do
       Bolt.Sips.transaction(main_conn, fn conn ->
-        book = Bolt.Sips.query(conn, "CREATE (x:XactCommit {foo: 'bar'}) return x")
-        assert {:ok, [row]} = book
-        assert row["x"].properties["foo"] == "bar"
+        assert {:ok, %Response{results: books}} =
+                 Bolt.Sips.query(conn, "CREATE (x:XactCommit {foo: 'bar'}) return x")
+
+        # TODO: maybe we can make Entity implement Access? That will avoid the Map gets below
+        assert "bar" ==
+                 books
+                 |> List.first()
+                 |> Map.get("x")
+                 |> Map.get(:properties)
+                 |> Map.get("foo")
 
         # Main connection should not see this new node.
-        {:ok, results} = Bolt.Sips.query(main_conn, "MATCH (x:XactCommit) RETURN x")
+        {:ok, %Response{results: results}} =
+          Bolt.Sips.query(main_conn, "MATCH (x:XactCommit) RETURN x")
+
         assert is_list(results)
 
         assert Enum.count(results) == 0,
@@ -76,7 +94,9 @@ defmodule Transaction.Test do
       end)
 
       # And we should see it now with the main connection.
-      {:ok, [%{"x" => node}]} = Bolt.Sips.query(main_conn, "MATCH (x:XactCommit) RETURN x")
+      {:ok, %Response{results: [%{"x" => node}]}} =
+        Bolt.Sips.query(main_conn, "MATCH (x:XactCommit) RETURN x")
+
       assert node.labels == ["XactCommit"]
       assert node.properties["foo"] == "bar"
     after